use simplified default logger
