+++ runhouse/resources/envs/conda_env.py
@@ -7,13 +7,14 @@ import yaml
 
 from runhouse.constants import ENVS_DIR
 from runhouse.globals import obj_store
+from runhouse.logger import get_logger
 
-from runhouse.logger import logger
 from runhouse.resources.envs.utils import install_conda, run_setup_command
-
 from runhouse.resources.packages import Package
 
 from .env import Env
+
+logger = get_logger(name=__name__)
 
 
 class CondaEnv(Env):
