+++ runhouse/resources/functions/function.py
@@ -1,16 +1,17 @@
 import inspect
-import logging
 from pathlib import Path
 from typing import Any, List, Optional, Tuple, Union
 
 from runhouse import globals
+from runhouse.logger import get_logger
 
-from runhouse.logger import logger
 from runhouse.resources.envs import Env
 from runhouse.resources.hardware import Cluster
 from runhouse.resources.module import Module
 
 from runhouse.resources.resource import Resource
+
+logger = get_logger(name=__name__)
 
 
 class Function(Module):
