+++ runhouse/resources/hardware/sky/command_runner.py
@@ -20,9 +20,9 @@ from runhouse.resources.hardware.sky import (
 )
 
 ##### RH modification #####
-import logging
-from runhouse.logger import logger
+from runhouse.logger import get_logger
 
+logger = get_logger(name=__name__)
 ##### RH modification #####
