+++ runhouse/resources/resource.py
@@ -5,8 +5,8 @@ from pathlib import Path
 from typing import Dict, List, Optional, Tuple, Union
 
 from runhouse.globals import obj_store, rns_client
+from runhouse.logger import get_logger
 
-from runhouse.logger import logger
 from runhouse.rns.top_level_rns_fns import (
     resolve_rns_path,
     save,
@@ -18,6 +18,8 @@ from runhouse.rns.utils.api import (
     ResourceAccess,
     ResourceVisibility,
 )
+
+logger = get_logger(name=__name__)
 
 
 class Resource:
