+++ runhouse/utils.py
@@ -31,9 +31,9 @@ from typing import Callable, Optional, Type, Union
 import pexpect
 
 from runhouse.constants import LOGS_DIR
+from runhouse.logger import get_logger
 
-logger = logging.getLogger(__name__)
-
+logger = get_logger(name=__name__)
 ####################################################################################################
 # Python package utilities
 ####################################################################################################
@@ -475,3 +475,66 @@ def generate_default_name(prefix: str = None, precision: str = "s", sep="_") -> 
     if prefix is None:
         return timestamp_key
     return f"{prefix}{sep}{timestamp_key}"
+
+
+####################################################################################################
+# Logger utils
+####################################################################################################
+class ColoredFormatter:
+    COLORS = {
+        "black": "\u001b[30m",
+        "red": "\u001b[31m",
+        "green": "\u001b[32m",
+        "yellow": "\u001b[33m",
+        "blue": "\u001b[34m",
+        "magenta": "\u001b[35m",
+        "cyan": "\u001b[36m",
+        "white": "\u001b[37m",
+        "reset": "\u001b[0m",
+    }
+
+    @classmethod
+    def get_color(cls, color: str):
+        return cls.COLORS.get(color, "")
+
+    # TODO: This method is a temp solution, until we'll update logging architecture. Remove once logging is cleaned up.
+    @classmethod
+    def format_log(cls, text):
+        ansi_escape = re.compile(r"(?:\x1B[@-_][0-?]*[ -/]*[@-~])")
+        return ansi_escape.sub("", text)
+
+
+class ClusterLogsFormatter:
+    def __init__(self, system):
+        self.system = system
+        self._display_title = False
+
+    def format(self, output_type):
+        from runhouse import Resource
+        from runhouse.servers.http.http_utils import OutputType
+
+        system_color = ColoredFormatter.get_color("cyan")
+        reset_color = ColoredFormatter.get_color("reset")
+
+        prettify_logs = output_type in [
+            OutputType.STDOUT,
+            OutputType.EXCEPTION,
+            OutputType.STDERR,
+        ]
+
+        if (
+            isinstance(self.system, Resource)
+            and prettify_logs
+            and not self._display_title
+        ):
+            # Display the system name before subsequent logs only once
+            system_name = self.system.name
+            dotted_line = "-" * len(system_name)
+            print(dotted_line)
+            print(f"{system_color}{system_name}{reset_color}")
+            print(dotted_line)
+
+            # Only display the system name once
+            self._display_title = True
+
+        return system_color, reset_color
