+++ tests/test_obj_store.py
@@ -5,6 +5,7 @@ import numpy as np
 import pytest
 
 import runhouse as rh
+from runhouse.logger import get_logger
 
 from tests.test_resources.test_modules.test_functions.test_function import (
     multiproc_np_sum,
@@ -13,8 +14,7 @@ from tests.test_resources.test_modules.test_functions.test_function import (
 TEMP_FILE = "my_file.txt"
 TEMP_FOLDER = "~/runhouse-tests"
 
-from runhouse.logger import logger
-
+logger = get_logger(name=__name__)
 
 UNIT = {"cluster": []}
 LOCAL = {
